Refactor scenario form components and hooks location
Moved CreateScenarioForms components, constants, and related hooks from the project directory to the scenario directory. Updated all relevant imports to reflect the new paths, improving code organization and separation of concerns.
